static page e cron secret

+++ src/app/api/restartList/route.ts
@@ -1,9 +1,16 @@
 import { deleteGamesList } from "@/utils/kv-games";
 import { updateCurrentGame } from "@/utils/update-current-club";
 import { revalidatePath } from "next/cache";
+import { NextRequest } from "next/server";
 
-export async function GET() {
-    await deleteGamesList();
+export async function GET(request: NextRequest) {
+    const authHeader = request.headers.get('authorization');
+	if (authHeader !== `Bearer ${process.env.CRON_SECRET}`) {
+		return new Response('Unauthorized', {
+			status: 401,
+		});
+	}
+	await deleteGamesList();
 	const result = await updateCurrentGame();
 	revalidatePath('/');
 	return Response.json(result)
